Improve code quality
Assistinator class handles execution of command. 0 and 1 are used to
represent if a task is not done or done respectively.

Multiple levels of abstraction in Assistinator class leading to
violation of SLAP. Arbitrary numbers 0 and 1 used to represent state of
tasks makes code not obvious.

Let's,
* extract all functions to execute command into a separate class
* create a enum class to represent states of tasks

CommandExecutor class helps to abstract away command execution logic so
that Assistinator focuses on higher level application logic. Using DONE
NOTDONE for task status is more intuitive.