Add assert calls to different methods
There are no calls to assert. This can cause the program to
experience error when assumptions do not hold true. This will
cause the program to crash without much clue to where the error
has taken place.

Adding in multiple assert lines allows us to check if the
assumptions made are true during run time. It also let us know
where the error was when the assumptions fails.

Let's add the assumptions made for different methods to be
checked using the assert method call.