Add assert statements in key parts of code
Some key assumptions, like that the list of tasks exists, and is not
null, are not tested.

By testing these key assumptions, although they may hold true for now,
when doing future refactoring or feature-adding, these assumptions may
break.

These new assert statements help to ensure that this is caught.