Add Java assertions to project
Command classes: add Java assertions

Java assertions helps to identify potential bugs within the code base by
stating assumptions about the logic of the code.

Java assertions are added at certain points within the function mainly
at the start of the function to ensure that the input is not null.

These assertions are placed at this specfic points is because if the
input is null it means that the program is not capturing and passing the
input correctly to the respective function.