Add assertions at key locations
Lack of areas to check when the code gets long.
It is common to miss out on bugs which can cause the program to fail.

Write assertions to important areas allows us to ensure our
assumptions is correct. For example, the tasklist file is created by the
code.

Using assertions ensure that the intended effect of the methods is done
correctly.