Add assertion to various methods
The data classes and exception classes assumes the arguments passed in
are valid. The validity is assumed the arguments passed to be not null
or blank.

These assertion act as a secondary checks for the various data classes
as the guard clauses should have caught such errors parsing the input.

Let's add null assertion and empty assertion checks to
* Data classes such as TodoItem, EventItem, DeadlineItem and TodoList.
* Exception classes such as TohruException and CorruptSaveException.