Improve coding quality
There was no coding standard in the project.
I have added a coding standard to the project to
improve the coding quality.

I followed the principles outlined in the Week 5
notes, such as to have a Single Level of Abstraction
and to reduce arrowhead occurences in the code.

This coding standard will be followed for the rest
of the project.