Update code to handle time for Deadline and Event classes
Currently, the chatbot only handles dates in very specific formats which are not natural to users.

Since we want users to find this chatbot intuitive and user friendly, it would be subpar to not have this feature.

Let's update the Parser class, Event class and Deadline class to allow for usage of more natural dates.