Adds assertions for ListByDateCommand and TaskList
There is no any assertion in the program.

These are added to test assumptions of the two methods.

Adds assertions to check if keyword is valid for findTasksByKeyword method to process and to check if date is blank for the execution of ListByDateCommand.

The date should not be empty, otherwise, the program will not find any task. The keyword for finding tasks should not be empty too.

This is different from exceptions, since the validation is already done before these methods can be called.