Remove scheduledDate attribute from Task
ToDo class, which inherits Task, does not need access to the scheduledDate attribute. Only ScheduledTask inheritors, Deadline and Event, will require access to scheduledDate.

This removes the need to explicitly check for Deadline or Event objects in viewScheduledTasks method in TaskList. Makes the code more readable and maintanable.

Let's remove the scheduledDate attribute from Task and add it into ScheduledTask.