Remove unused imports in TaskList class and change assertions in Parser class
Previously assertions in Parser class are used wrongly to catch exceptions which led to build failure as test cases in the test file were failing.

This commit will delete the assertions in the Parser class as these exceptions are handled explicitly by the try catch block in the Parser class so there is no need for assert statements here.