Add Tagging Functionalities
There is currently no way to group similar tasks together, and the task
list remains sorted only based on order of addition.

There must be a method to group tasks so the find command can
be put to better use.

Tagging functionality has been added so that similar tasks can be
found together in the same list by searching the tag.

This is preferable so that different tasks added at different times but
are related can be viewed together.