Refactor Command classes
The MarkCommand, UnmarkCommand and DeleteCommand classes has duplicate
code chunks that deals with parsing the argument passed in as an
integer. The ListCommand and FindCommand classes has duplicate code
chunks that deals with converting to-do entries into itemised string
format. Additionally, The AddDeadlineCommand and AddEventCommand
classes does quite a few checks on data validity, which makes the
method feel bloated.

These duplicated code makes it a hassle to update the logic for the
argument parsing in each file for the above Command classes.

Let's refactor
* MarkCommand, UnmarkCommand and DeleteCommand classes to use a
common integer parser.
* ListCommand and FindCommand classes to use a common
getItemisedItemsString() method to get the itemised string
representation.
* AddDeadlineCommand and AddEventCommand to use a separate
method to handle the parsing of arguments.