Supports undo command
Users are not allowed to undo a command.

An undo command gives users flexibility to make mistakes.

A new UndoCommand class and Undoable interface are added. A stack is added to keep track of those undoable commands.

Using stack maintains the order of commands to be undone due to its first-in-first-out property.