A-Personality: Add personality quirk to Jackson
Jackson gives standard responses and behaves like a normal chatbot.

This can seem mundane and uninteresting to users and they may not
enjoy or remain engaged with it.

Let's add a "Singlish" style response that Jackson can use to
add flavour to the program.

This is a placeholder commit as I have already integrated this
personality many commits ago.