A-MoreErrorHandling: Improve error handling part 2
Jackson does not have the capability to detect conflicting task
names before adding the tasks.

The possibility of a name conflict impeding code progress later on
is high and minimising this risk is optimal for now.

Let's,
* create DuplicatedTaskException to handle such conflicts
* modify addTask in TaskList.java to check for name conflicts first
* modify Javadocs to adhere to coding standard and checkStyle

I forgot to stage the changes so this commit will be an extension
to the previous one.