Add assert for TaskList class
There are several methods which manipulate the TaskList and its
contents depending on the method parameters. As such, it is
crucial that the parameters are not invalid, as this logic is
handled by the programmer.

Including asserts will help guard against programmer's errors of
passing in invalid parameters to these methods.