Move code on format list in Ui.java to elsewhere
formatToList(), formatToFullList(), formatToMatchedList() are private
methods in Ui class. But the Ui is to show the output of the GUI bot,
not meant to further modify the task list format. Such further
processing should be done during the execution of the command.

This is to organise the code better.

Let's
* Create FormatListCommand class which extends Command class
* Add formatToList() function in FormatListCommand class as a protected
  method
* Extend ListCommand, FindCommand class from FormatListCommand class
* Delete formatToFullList() and formatToMatchedList()

FormatListCommand class represents Commands that require formatting the
list of tasks and output the formatted list. All these commands will
need to use the formatToList(), so this function should be protected so
that commands belonging to FormatListCommand() can use this method.
formatToFullList() and formatToMatchedList() are pretty similar to
formatToList(), thus there is no need to extract it (although
technically the naming is more representative).