Move formatToList() code as a method of TaskList
formatToList(), formatToFileList() are placed in FormatListCommand and
Storage class respectively. But these methods can be a characteristic of
the taskList object representation, similar to how it is done for Task
objects. Moreover, code can be written in a more declarative style by
making use of Streams.

Let's
* Move formatToList() and formatToFileList() as methods for TaskList
  class
* Rewrite in declarative style, for loops to traverse through the list
  can be rewritten to make use of streams

This is to enhance code readability and better organisation of code.