Add assert to Parser, TaskList, Storage and Command child classes.
We did not check whether the format of the user inputs are correct.
The user may incorrectly input their commands as they are not aware of
the format which leads to unnecessary errors. To fix this, add
assertions. This checks if the user input for specific instructions
has the correct format and does not raise unnecessary errors like
IndexOutOfBoundsException.

Add try-catch for DateTimeParseException in Deadline and Events class.

It is not clear to the user which datetime format to use. There are
many different datetime formats such as MM/DD/YYYY or DD/MM/YYY.
Therefore, standardising the format to YYYY-MM-DD would make the
chatbot easier to use.