Change push to add for methods adding into I/O buffer in UiGui
All statements that adds Strings into input and output buffer pushes
them in the wrong order.

When polled, the buffer returns the queued strings in the wrong order,
as seen with `find` command - the list printout comes before the
confirmation reply.

All instances of `xBuffer.push` to be replaced with `xBuffer.add`
(where x is `input` or `output`).

This ensures the Strings are properly enqueued into the buffer to be
polled in order.

Note: ensure push is paired with pop, and add is paired with poll for
future uses of ArrayDeque.