Improve code quality
To make the code more readable, code quality must be improved. This is to ensure that someone reading the code is able to understand what is going on at all times.

Let's edit any lines of code that has deep nesting to code without, avoid ambiguous terms, and add comments before each method or class to explain its purpose.