Remove the redundant regex in CommandManager enum
Previously, there was a field the CommandType enum called regex that
required you to specify the regex for the command.

However, this can be done dynamically based on the arguments of the
commands that are present.

Hence, removed the regex field and dynamically create the regex based
on the argument list.

This helps to reduce code burden in needing to specify the regex, which
might get into errors