Add tag command to Quack
Tag Command: allows assigning of tags to tasks

Allow rhe user to better organize their tasks.

Add a new command tag which will prompt the user for an index and a tag
label and then add the label to the task.
There us also an untag command which will prompt the user for an
index and then remove the tag from the task.

Asking the user for the index of where the task is located will make
finding easier. And as for prompting the user for a tag is to allow
freedom to choose the taging system the user wants to impliment.