Add assertions to various points in the code
There are no assertions in the code.

Assertions can be used to define assumptions about the program state
so that the runtime can verify them.

Let's
* add assertions to Sentinel methods to check if task list exists
* add assertions to Storage file load() method to check if the file
was created correctly
* add assertions to TaskList methods to check if the task to modify
exists

Assertions are used this way to verify that the assumptions made
about the program are correct and to catch potential bugs.

See
https://nus-cs2103-ay2425s1.github.io/website/schedule/
week5/topics.html#w5-6-assertions