ListCommandTest class: Modify to fit the GUI implementation
The runCommand_nonEmptyList_success() test case in ListCommandTest class
fails.

The failure is caused by its implementation, which involves reading
output to the terminal. But since the current implementation is GUI,
there are no more outputs to the terminal.

Let's modify the method to read the returning string from the
runCommand() method for testing.

Doing this is preferrable as the string returned from runCommand() is
being passed and subsequently displayed on the GUI, so it is the same
one as the one seen from the GUI, ensuring that it is a valid test case.