Improve code quality
Some comments are unnecessary and repeat information already obvious
from the code. Some comments also explain the "how" of the code, and
not the "what" and "why".

This reduces the code quality of my code.

Let's improve the naming of variables and methods, avoid
complicated expressions and remove unnecessary comments.

Improving the names of variables and code structure allows my code to
be more self-explanatory thus reducing the need for unnecessary
comments.