When command list is entered, only tasks entered in that respective run are returned, not all those stored in Data file
The feature of being able to store the tasks data and refer to them again is not being used, hence defeats the purpose of the chatbot being able to list all the tasks.

New method in Storage class that returns the String form of all tasks stored in Data file.

A String is returned instead of the usual printing of content directly, such that the String can be used to work with the JavaFx files as well.