Modify Storage.java and StorageTest.java
Originally Storage.java only work on a default file "data/Kitty.txt",
thus both the Chatbot and StorageTest work on this file. Modify the
Storage.java to ensure that it works on the file specified during the
initialization. This adds generality to the Storage class and
helps to separate the Chatbot and Test.