Create guard case for list command
There is no special response message when the 'list' command is
input with an empty task list.

Creating a special response message makes the chatbot response more
realistic.

Let's create a guard case using an if statement to return a special
response string when 'list' is input with an empty task list.