Add assertions to the main logic flow
Assertions are not added yet, thus we are not able
to catch any illogical flow in the program.

Let's add assertions to parts of the program
where it is important for the program to run
smoothly.

In the run() method, we assert that the ui is not null.
In the getResponse method, we assert that the input
is not null and instead a String.
In the saveTasks method in the Storage class, we assert
that taskList is not null.