Add assertions in the code base
Add assertion at 2 place in Storage.java file to
check if the file exists after try to create the
database at load() function and before try to save
data into the database at saveData() function.

As a result, we are sure that the file always
exists when we access it.