Add assertion
Command should be a String for the programme to work. Or else, the methods that are going to be called will not work.

Let's add an assertion to check if variable "command" is a String.

Adding an assertion will ensure that the command variable has a String, preventing any potential problems in the future.