MainGUI.java: Add scroll to bottom when new message is printed
Previously, the UI does not follow the new messages entered by the user
and the response that is printed by the chatbot.

This commit adds a listener to the height of the dialogContainer, to
scroll to the bottom of the page whenever it changes.

This improves the user experience and makes the app more similar to a
typical messaging app that users would be familiar with.