Skibidi.java: Add logic to update file every command
The tasks.txt file is not updated when running the chatbot
with javafx.

Move the saveTasksToDisk method from within the start method of
Skibidi.java to the getResponse method, which executes everytime a
command is entered by the user.

This automatically keeps the tasks.txt file up to date with the latest
task list.