Add detection of duplicates of task
User should not be able to add duplicated task into the list (regardless
the task in the list is marked done or undone).

Application now is able to detect such duplicates. Exception will be
thrown if user trying to add duplicated task.

Another comparing method is added to Task objects to not compare status
unlike the overridden equals method.