Change order of assignment in Lexi:getResponse
The program return a NullPointerException when list command
is executed when the data file is empty as ListCommand:execute
throws a LexiException when the list of tasks are empty.

As a result, the commandType assignment does not execute as
the program exits the code and reaches the catch block. This
later causes NullPointerException in DialogBox:changeDialogStyle
as commandType is null. It is null as the program exited the
try block before the assignment statement.

Let's perform the assignment of the commandType before executing
the command itself.