Task list no longer allows user to add duplicate tasks
Task list allows user to add duplicate tasks.

Allowing users to add duplicate tasks introduces clutter into the
list and makes it harder for the user to read.

Let's store the task descriptions of newly added tasks into a map
so that we can detect when a duplicate task (i.e. same description
and/or date) has been added.

Using a map is preferable as it has O(1) lookup time.