Event tasks: add checking of schedule clashes
Users can add events even if there are clashes in time periods with
other existing events.

Users will not be aware of clashing events and will have to deal with
conflicting events later on.

Let's check for conflicts when adding new events with existing events.
If there are conflicts, warn the users which existing event conflicts
with the new event.

This informs users of clashes in schedule and the user can then take
actions to resolve the clashes.