Change implementation of getTasksWithTerm(String) method
Method getTasksWithTerm uses for loop to iterate and search for
tasks.

Use of for loop is verbose and difficult to understand.

Let's switch to using streams to cut down and clarify what the code
is doing. Streams has many predefined functions which we can use to
cut down on code verbosity.