Parser.java: Deal with duplicate items
The chatbot allows multiple tasks with the same name to be added.

This causes confusion for the user who may accidentally add multiples
of one task.

Let's modify addToDo, addDeadline, addEvent to not add a task if a task
with the same name already exists.

This will prevent the user from making a copy of an existing task, and
notify them about it as well, reducing potential confusion.