Fix could not add task if name was part of existing task name
Last commit uses old TaskList.find method to check if the task name
already exists.

This is wrong as the old names only needed to contain the new name to
be counted as duplicate (e.g. "SoC career fair" contains "career fair")

Let's add "exact" param to TaskList.find to set whether we are looking
for an exact match.

Results in adding new tasks performing as expected (can add as long as
name is not exact duplicate)