Add Assertions to taskList related methods
Sometimes we may want to check if the method that calls taskList method
successfully manipulates the list inside the taskList.

In some cases programmer may forget to add new Task to the list when
writing methods for a new type of class.

Adding assertions to these methods can
verify assumptions about Internal
Invariants, Control-Flow Invariants, Preconditions, Post conditions, and
Class Invariants.

Assertions will remind the possible flaw in the terminal if it's asserted
to be false.