Delete command: fix tasks not saving after delete
Delete command deletes from TaskList but does not update the local .txt file.
If the user deletes and exits without another command to save,
the deletion is not saved. This is inconsistent with other commands
that save the TaskList after successful execution.

Let's save TaskList to local file after every successful deletion.