Refactor code
Long methods and complicated expressions are used to parse through each user command in the Parser class.

Long methods and complicated expressions make it hard for the reader to understand the coding logic.

Breaking down long methods and complicated expressions into smaller methods make their purposes clearer and more understandable for the reader.

Let's break down the matching and execution of each command into different methods.

This makes it easier for the reader to recognise how each command is recognised and matched, as well as how each command is executed.