Add Priority To Task
To give the user flexibility the adjust the priority of the task

Add the "set" command in Jay chatbot to set the priority of task,
change the format of the Task string, and also the Task storage string
format. Make the MainWindow show the tasks with high priority when the
chatbot start. And also modify the test cases.