refactor: Add assertions to improve code robustness
Task index that are passed into the task list should be in a valid
range. (i.e. from 0 to number of tasks - 1)

This is already checked for in the various action classes. However,
using assertions to ensure this would make the code more robust.

Let's add assertions for the mark, unmark and delete methods in
the TaskList class to do so.