Extract out methods within writeFile() in Storage class
The writeFile() method in the Storage class was too long with multiple
switch cases to write to the file in the case of the different tasks.

Extracting out a method to get the format in which the task should
be written to the file improves the code quality.

Let's extract the switch case code block into a method in the respective
classes to return a String in the proper format to be written to the
task file.