refactor: Abstract out several string constants
There are several fixed messages that BotManager has (for example welcome and goodbye messages). They currently appear as magic strings in the code.

While the current implementation is fine for now, it may cause problems if these strings need to be reused in other parts of the codebase in the future.

Let's abstract these strings out into as enum class as constants.