Add implementation to support B-recurringTasks
A new abstract class "RecurringTypeTask" was created to add the recurringType parameter to all tasks
that can have recurrences. Deadline and Event classes now extend from RecurringTypeTask

An enum is used to define the recurring type of tasks, with the following possible options:
* Daily
* Weekly
* Monthly
* Bimonthly
* Quarterly
* Biannually
* Annually

The format which tasks are saved in in the save file is also changed to accomodate the addition of the
recurrence type.

Upon launching the application Slave will:
* Highlight the tasks that have ended prior to the day of launching the application.
This list of tasks will not be stored after closing the application.
* Create new recurring tasks for tasks that have ended prior to the day of launching the application
and add them to the task list.
* Remove all tasks that have ended prior to the day of launching the application.

Slave will allow the creation of tasks that have already ended prior to the day of creation, to allow
users to not have to calculate date of its next occurance for recurring tasks.
For non recurring tasks, it does not affect the functionality of the application as it will be removed
from the task list the next time the application is run