Improve code quality
The code suffers from complicated expressions and poor ordering of
statements as well as a lack of code comments.

These make it hard to read and understand what the code is doing
at certain points.

Let's simplify some code expressions, put the guard clause in front
to make the happy path prominent and add some comments to explain
what the code is doing to enhance overall readability of the code.