Refactor code to improve code quality
Some methods are longer than 30 LoC, minor presence of arrowhead
style code and weak SLAP.

These factors decrease the readability of the code, in turn
reducing the code quality.

Let's
* use Extract Method to reduce the LoC of some methods in Sentinel
class and to increase SLAP.
* split parseStringToTask in Parser class into several methods to
reduce the length of the method and to increase SLAP.
* extract out a loadFile method from Storage.load() to reduce the
overall length of the method.